third_party/WebKit/LayoutTests/fast/html/imports/rel-style-to-import.html

<!DOCTYPE html>

<script>
    jsTestIsAsync = true;
</script>

<script src="../../../resources/js-test.js"></script>
<script src="resources/import-helpers.js"></script>

<link id="target" rel="stylesheet" href="resources/hello.css">

<script>
debug("Updating attribute value shouldn't result import loading.");
debug("Before update:");
shouldBeNonNull("target.sheet");
shouldBeNull("target.import");
target.setAttribute("href", "resources/hello.html");
target.setAttribute("rel", "import");
var targetLoaded = new Promise(function(resolve) {
    target.onload = function() {
        resolve();
    };
});
var placeholder = createPlaceholder();
var placeholderLoaded = new Promise(function(resolve) {
    placeholder.onload = function() {
        resolve();
    };
});
Promise.all([placeholderLoaded, targetLoaded]).then(function() {
    debug("After update:");
    shouldBeNonNull("target.sheet");
    shouldBeNull("target.import");
    finishJSTest();
});
</script>
