third_party/WebKit/LayoutTests/fast/html/imports/rel-style-to-import-expected.txt

CONSOLE WARNING: CSS cannot be loaded from `file:` URLs unless they end in a `.css` file extension.
Updating attribute value shouldn't result import loading.
Before update:
PASS target.sheet is non-null.
PASS target.import is null
After update:
PASS target.sheet is non-null.
PASS target.import is null
PASS successfullyParsed is true

TEST COMPLETE
